Restore original body of issue/pull request
This is related to issue #674.  Before this commit, skipping the editor
returned an empty body in the created issue or pull request.

It seems that because the editor was never opened, a temporary file is
never created and therefore the commit messages aren't appended.  This
leads to returning an empty body.

This commit restores the original body _before_ we initiated the prompt
to edit.